# _15_appliance_store/api/store/v1/supplier_router.pb.go
// Code generated by https://github.com/go-dev-frame/sponge, DO NOT EDIT.

package v1

import (
	"context"
	"errors"
	"strings"

	"github.com/gin-gonic/gin"
	"go.uber.org/zap"

	"github.com/go-dev-frame/sponge/pkg/errcode"
	"github.com/go-dev-frame/sponge/pkg/gin/middleware"
)

type SupplierLogicer interface {
	Create(ctx context.Context, req *CreateSupplierRequest) (*CreateSupplierReply, error)
	DeleteByID(ctx context.Context, req *DeleteSupplierByIDRequest) (*DeleteSupplierByIDReply, error)
	UpdateByID(ctx context.Context, req *UpdateSupplierByIDRequest) (*UpdateSupplierByIDReply, error)
	GetByID(ctx context.Context, req *GetSupplierByIDRequest) (*GetSupplierByIDReply, error)
	List(ctx context.Context, req *ListSupplierRequest) (*ListSupplierReply, error)
}

type SupplierOption func(*supplierOptions)

type supplierOptions struct {
	isFromRPC  bool
	responser  errcode.Responser
	zapLog     *zap.Logger
	httpErrors []*errcode.Error
	rpcStatus  []*errcode.RPCStatus
	wrapCtxFn  func(c *gin.Context) context.Context
}

func (o *supplierOptions) apply(opts ...SupplierOption) {
	for _, opt := range opts {
		opt(o)
	}
}

func WithSupplierHTTPResponse() SupplierOption {
	return func(o *supplierOptions) {
		o.isFromRPC = false
	}
}

func WithSupplierRPCResponse() SupplierOption {
	return func(o *supplierOptions) {
		o.isFromRPC = true
	}
}

func WithSupplierResponser(responser errcode.Responser) SupplierOption {
	return func(o *supplierOptions) {
		o.responser = responser
	}
}

func WithSupplierLogger(zapLog *zap.Logger) SupplierOption {
	return func(o *supplierOptions) {
		o.zapLog = zapLog
	}
}

func WithSupplierErrorToHTTPCode(e ...*errcode.Error) SupplierOption {
	return func(o *supplierOptions) {
		o.httpErrors = e
	}
}

func WithSupplierRPCStatusToHTTPCode(s ...*errcode.RPCStatus) SupplierOption {
	return func(o *supplierOptions) {
		o.rpcStatus = s
	}
}

func WithSupplierWrapCtx(wrapCtxFn func(c *gin.Context) context.Context) SupplierOption {
	return func(o *supplierOptions) {
		o.wrapCtxFn = wrapCtxFn
	}
}

func RegisterSupplierRouter(
	iRouter gin.IRouter,
	groupPathMiddlewares map[string][]gin.HandlerFunc,
	singlePathMiddlewares map[string][]gin.HandlerFunc,
	iLogic SupplierLogicer,
	opts ...SupplierOption) {

	o := &supplierOptions{}
	o.apply(opts...)

	if o.responser == nil {
		o.responser = errcode.NewResponser(o.isFromRPC, o.httpErrors, o.rpcStatus)
	}
	if o.zapLog == nil {
		o.zapLog, _ = zap.NewProduction()
	}

	r := &supplierRouter{
		iRouter:               iRouter,
		groupPathMiddlewares:  groupPathMiddlewares,
		singlePathMiddlewares: singlePathMiddlewares,
		iLogic:                iLogic,
		iResponse:             o.responser,
		zapLog:                o.zapLog,
		wrapCtxFn:             o.wrapCtxFn,
	}
	r.register()
}

type supplierRouter struct {
	iRouter               gin.IRouter
	groupPathMiddlewares  map[string][]gin.HandlerFunc
	singlePathMiddlewares map[string][]gin.HandlerFunc
	iLogic                SupplierLogicer
	iResponse             errcode.Responser
	zapLog                *zap.Logger
	wrapCtxFn             func(c *gin.Context) context.Context
}

func (r *supplierRouter) register() {
	r.iRouter.Handle("POST", "/api/v1/supplier", r.withMiddleware("POST", "/api/v1/supplier", r.Create_48)...)
	r.iRouter.Handle("DELETE", "/api/v1/supplier/:id", r.withMiddleware("DELETE", "/api/v1/supplier/:id", r.DeleteByID_46)...)
	r.iRouter.Handle("PUT", "/api/v1/supplier/:id", r.withMiddleware("PUT", "/api/v1/supplier/:id", r.UpdateByID_46)...)
	r.iRouter.Handle("GET", "/api/v1/supplier/:id", r.withMiddleware("GET", "/api/v1/supplier/:id", r.GetByID_46)...)
	r.iRouter.Handle("POST", "/api/v1/supplier/list", r.withMiddleware("POST", "/api/v1/supplier/list", r.List_48)...)

}

func (r *supplierRouter) withMiddleware(method string, path string, fn gin.HandlerFunc) []gin.HandlerFunc {
	handlerFns := []gin.HandlerFunc{}

	// determine if a route group is hit or miss, left prefix rule
	for groupPath, fns := range r.groupPathMiddlewares {
		if groupPath == "" || groupPath == "/" {
			handlerFns = append(handlerFns, fns...)
			continue
		}
		size := len(groupPath)
		if len(path) < size {
			continue
		}
		if groupPath == path[:size] {
			handlerFns = append(handlerFns, fns...)
		}
	}

	// determine if a single route has been hit
	key := strings.ToUpper(method) + "->" + path
	if fns, ok := r.singlePathMiddlewares[key]; ok {
		handlerFns = append(handlerFns, fns...)
	}

	return append(handlerFns, fn)
}

func (r *supplierRouter) Create_48(c *gin.Context) {
	req := &CreateSupplierRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.Create(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *supplierRouter) DeleteByID_46(c *gin.Context) {
	req := &DeleteSupplierByIDRequest{}
	var err error

	if err = c.ShouldBindUri(req); err != nil {
		r.zapLog.Warn("ShouldBindUri error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	if err = c.ShouldBindQuery(req); err != nil {
		r.zapLog.Warn("ShouldBindQuery error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.DeleteByID(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *supplierRouter) UpdateByID_46(c *gin.Context) {
	req := &UpdateSupplierByIDRequest{}
	var err error

	if err = c.ShouldBindUri(req); err != nil {
		r.zapLog.Warn("ShouldBindUri error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.UpdateByID(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *supplierRouter) GetByID_46(c *gin.Context) {
	req := &GetSupplierByIDRequest{}
	var err error

	if err = c.ShouldBindUri(req); err != nil {
		r.zapLog.Warn("ShouldBindUri error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	if err = c.ShouldBindQuery(req); err != nil {
		r.zapLog.Warn("ShouldBindQuery error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.GetByID(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *supplierRouter) List_48(c *gin.Context) {
	req := &ListSupplierRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.List(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}